(FACT-1716) Collect OEM strings in DMI source
Adds a vector of OEM strings to the collected data from the DMI source.
This information is not available in the /sys/class/dmi/id/ directory,
so it can only be collected with root privileges via dmidecode (or the
DMI table).

Not all vendors store useful information in the OEM strings, but
VirtualBox, for example, reports the hypervisor version and revision
there.

Also refactors DMI tests so that a single fixture class can use fixtures
for files in /sys/class/dmi/id/ and dmidecode output simultaneously.