(maint) Refactor DMI data source
Previously, the DMI data source collected data in its constructor, so
even when DMI wouldn't have been used anyway, it would still try to
collect its data from dmidecode and /sys on instantiation. This commit
changes the DMI source so that it only collects data when one of data
values is accessed.