(maint) Only preserve ebx when it's absolutely necessary
Preserving ebx is only needed on i386 when building position-
independent code, since it is permanently reserved for PIC
purposes, and is invisible to the compiler's register allocator.

On x86_64 and on non-PIC x86, we can rely on the compiler to
properly handle our clobbering of that register. This is a
compiler bug in my view, but that doesn't stop us from having
to deal with it.